pyproject.toml
[project]
name = "mcp-copilot-vscode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]==1.10.1",
    "python-dotenv>=1.1.1",
]
